+++ Runtime/AdvancedDropdownList.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using UnityEngine;
 
 namespace SaintsField
 {
@@ -8,20 +9,20 @@ namespace SaintsField
         {
         }
 
-        public AdvancedDropdownList(string displayName, bool disabled = false, string icon = null)
-            : base(displayName, disabled, icon)
+        public AdvancedDropdownList(string displayName, bool disabled = false, string icon = null, Color? color = null)
+            : base(displayName, disabled, icon, color)
         {
         }
 
         public AdvancedDropdownList(string displayName, T value, bool disabled = false, string icon = null,
-            bool isSeparator = false)
-            : base(displayName, value, disabled, icon, isSeparator)
+            Color? color = null, bool isSeparator = false)
+            : base(displayName, value, disabled, icon, color, isSeparator)
         {
         }
 
         public AdvancedDropdownList(string displayName, IEnumerable<Dropdown<T>> children, bool disabled = false,
-            string icon = null, bool isSeparator = false)
-            : base(displayName, children, disabled, icon, isSeparator)
+            string icon = null, Color? color = null, bool isSeparator = false)
+            : base(displayName, children, disabled, icon, color, isSeparator)
         {
         }
     }